extract handleButtonPress
