+++ src/closeOpenParens.js
@@ -1,6 +1,6 @@
 const closeOpenParens = (result) => {
-  var numberOfOpenP = (result.match(/\(/g) || []).length;
-  var numberOfCloseP = (result.match(/\)/g) || []).length;
+  var numberOfOpenP = ((result + "").match(/\(/g) || []).length;
+  var numberOfCloseP = ((result + "").match(/\)/g) || []).length;
 
   while (numberOfOpenP > numberOfCloseP) {
     result = result + ")";
